+++ green-eggs-app/package.json
@@ -10,13 +10,17 @@
     "web": "expo start --web"
   },
   "dependencies": {
+    "@react-navigation/native": "^6.1.1",
+    "@react-navigation/stack": "^6.3.10",
     "expo": "~47.0.8",
     "expo-status-bar": "~1.4.2",
     "react": "18.1.0",
     "react-native": "0.70.5",
+    "react-native-gesture-handler": "~2.8.0",
     "react-native-paper": "^5.0.2",
     "react-native-reanimated": "^2.13.0",
-    "react-native-safe-area-context": "4.4.1"
+    "react-native-safe-area-context": "4.4.1",
+    "react-native-screens": "~3.18.0"
   },
   "devDependencies": {
     "@babel/core": "^7.12.9"
