+++ green-eggs-app/pages/ProfilePage.jsx
@@ -1,0 +1,12 @@
+import { View, Text } from 'react-native'
+import React from 'react'
+
+const ProfilePage = () => {
+  return (
+    <View>
+      <Text>My Profile</Text>
+    </View>
+  )
+}
+
+export default ProfilePage
